reduce sideloads on cluster list
