+++ assets/src/components/cd/clusters/runtime/RuntimeServices.tsx
@@ -1,13 +1,10 @@
 import { Table } from '@pluralsh/design-system'
-import { ClustersRowFragment, useRuntimeServicesQuery } from 'generated/graphql'
-import { Key, useMemo, useRef, useState } from 'react'
-import { useTheme } from 'styled-components'
+import { ClustersRowFragment, RuntimeServicesQuery } from 'generated/graphql'
+import { useMemo } from 'react'
 
 import { isNonNullable } from 'utils/isNonNullable'
 
 import { runtimeColumns } from './columns'
-
-const POLL_INTERVAL = 10 * 1000
 
 export function getClusterKubeVersion(
   cluster: Nullable<Pick<ClustersRowFragment, 'currentVersion' | 'version'>>
@@ -16,23 +13,10 @@ export function getClusterKubeVersion(
 }
 
 export default function RuntimeServices({
-  cluster,
+  data,
 }: {
-  cluster?: ClustersRowFragment | undefined | null
+  data?: RuntimeServicesQuery
 }) {
-  const theme = useTheme()
-  const kubeVersion = getClusterKubeVersion(cluster)
-  const tabStateRef = useRef<any>()
-  const [tabKey, setTabKey] = useState<Key>('blocking')
-  const { data } = useRuntimeServicesQuery({
-    variables: {
-      kubeVersion,
-      hasKubeVersion: true,
-      id: cluster?.id ?? '',
-    },
-    fetchPolicy: 'cache-and-network',
-    pollInterval: POLL_INTERVAL,
-  })
   const addOns = useMemo(
     () => data?.cluster?.runtimeServices?.filter(isNonNullable) || [],
     [data?.cluster?.runtimeServices]
@@ -42,16 +26,14 @@ export default function RuntimeServices({
     return <p style={{ marginLeft: '1rem' }}>No Add-Ons Detected</p>
 
   return (
-    <>
-      <Table
-        data={addOns}
-        columns={runtimeColumns}
-        reactTableOptions={{ meta: { clusterId: cluster?.id } }}
-        css={{
-          maxHeight: 258,
-          height: '100%',
-        }}
-      />
-    </>
+    <Table
+      data={addOns}
+      columns={runtimeColumns}
+      reactTableOptions={{ meta: { clusterId: data?.cluster?.id } }}
+      css={{
+        maxHeight: 258,
+        height: '100%',
+      }}
+    />
   )
 }
