+++ assets/src/graph/cdClusters.graphql
@@ -118,9 +118,6 @@ fragment AddonVersionBlocking on AddonVersion {
 }
 
 fragment ClustersRow on Cluster {
-  apiDeprecations {
-    ...ApiDeprecation
-  }
   currentVersion
   id
   self
@@ -380,6 +377,9 @@ query RuntimeServices(
     version
     runtimeServices {
       ...RuntimeService
+    }
+    apiDeprecations {
+      ...ApiDeprecation
     }
   }
 }
